move SourceProvider to scaladsl package
