+++ akka-projection-core/src/main/scala/akka/projection/scaladsl/SourceProvider.scala
@@ -1,8 +1,4 @@
-/*
- * Copyright (C) 2020 Lightbend Inc. <https://www.lightbend.com>
- */
-
-package akka.projection
+package akka.projection.scaladsl
 
 import akka.stream.scaladsl.Source
