+++ akka-projection-slick/src/main/scala/akka/projection/slick/SlickProjection.scala
@@ -5,14 +5,15 @@
 package akka.projection.slick
 
 import scala.concurrent.duration.FiniteDuration
-
 import akka.Done
 import akka.annotation.ApiMayChange
-import akka.projection.{ Projection, ProjectionId, SourceProvider }
+import akka.projection.scaladsl.SourceProvider
+import akka.projection.{ Projection, ProjectionId }
 import akka.projection.slick.internal.SlickProjectionImpl
 import slick.basic.DatabaseConfig
 import slick.dbio.DBIO
 import slick.jdbc.JdbcProfile
+
 import scala.reflect.ClassTag
 
 @ApiMayChange
