+++ akka-projection-slick/src/main/scala/akka/projection/slick/internal/SlickProjectionImpl.scala
@@ -7,17 +7,18 @@ package akka.projection.slick.internal
 import java.util.concurrent.atomic.AtomicBoolean
 
 import scala.concurrent.duration.FiniteDuration
-
 import akka.Done
 import akka.actor.ClassicActorSystemProvider
 import akka.annotation.InternalApi
 import akka.event.Logging
-import akka.projection.{ Projection, ProjectionId, SourceProvider }
+import akka.projection.scaladsl.SourceProvider
+import akka.projection.{ Projection, ProjectionId }
 import akka.stream.KillSwitches
 import akka.stream.scaladsl.{ Flow, Sink, Source }
 import slick.basic.DatabaseConfig
 import slick.dbio.DBIO
 import slick.jdbc.JdbcProfile
+
 import scala.concurrent.{ ExecutionContext, Future, Promise }
 @InternalApi
 private[projection] object SlickProjectionImpl {
